ai-services/requirements.txt
# NusantaraExport.AI — AI Services Dependencies
# ================================================

# --- Framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# --- ASR (Whisper) ---
openai-whisper==20240930
soundfile==0.13.1

# --- LLM & NLP ---
transformers==4.47.1
torch>=2.1.0
sentence-transformers==3.3.1

# --- Market Gap Scoring ---
scikit-learn==1.6.1
xgboost==2.1.3
pandas==2.2.3
numpy==1.26.4

# --- Database ---
psycopg2-binary==2.9.10
pgvector==0.3.6

# --- Utilities ---
pydantic==2.10.4
python-dotenv==1.0.1
